PHPStan (Level 1) on controllers/

+++ controllers/admin/AdminCarrierWizardController.php
@@ -117,7 +117,11 @@ class AdminCarrierWizardControllerCore extends AdminController
             $carrier = new Carrier();
         }
 
-        if ((!$this->access('edit') && Tools::getValue('id_carrier')) || (!$this->access('add') && !Tools::getValue('id_carrier'))) {
+        if (
+            (!$this->access('edit') && Tools::getValue('id_carrier'))
+            || (!$this->access('add') && !Tools::getValue('id_carrier'))
+            || !isset($carrier)
+        ) {
             $this->errors[] = $this->trans('You do not have permission to use this wizard.', [], 'Admin.Shipping.Notification');
 
             return;
@@ -821,52 +825,54 @@ class AdminCarrierWizardControllerCore extends AdminController
                 }
             }
 
-            if ($carrier->is_free) {
-                //if carrier is free delete shipping cost
-                $carrier->deleteDeliveryPrice('range_weight');
-                $carrier->deleteDeliveryPrice('range_price');
-            }
-
-            if (Validate::isLoadedObject($carrier)) {
-                if (!$this->changeGroups((int) $carrier->id)) {
-                    $return['has_error'] = true;
-                    $return['errors'][] = $this->trans('An error occurred while saving carrier groups.', [], 'Admin.Shipping.Notification');
+            if (isset($carrier)) {
+                if ($carrier->is_free) {
+                    //if carrier is free delete shipping cost
+                    $carrier->deleteDeliveryPrice('range_weight');
+                    $carrier->deleteDeliveryPrice('range_price');
                 }
 
-                if (!$this->changeZones((int) $carrier->id)) {
-                    $return['has_error'] = true;
-                    $return['errors'][] = $this->trans('An error occurred while saving carrier zones.', [], 'Admin.Shipping.Notification');
-                }
-
-                if (!$carrier->is_free) {
-                    if (!$this->processRanges((int) $carrier->id)) {
+                if (Validate::isLoadedObject($carrier)) {
+                    if (!$this->changeGroups((int) $carrier->id)) {
                         $return['has_error'] = true;
-                        $return['errors'][] = $this->trans('An error occurred while saving carrier ranges.', [], 'Admin.Shipping.Notification');
+                        $return['errors'][] = $this->trans('An error occurred while saving carrier groups.', [], 'Admin.Shipping.Notification');
                     }
-                }
 
-                if (Shop::isFeatureActive() && !$this->updateAssoShop((int) $carrier->id)) {
-                    $return['has_error'] = true;
-                    $return['errors'][] = $this->trans('An error occurred while saving associations of shops.', [], 'Admin.Shipping.Notification');
-                }
+                    if (!$this->changeZones((int) $carrier->id)) {
+                        $return['has_error'] = true;
+                        $return['errors'][] = $this->trans('An error occurred while saving carrier zones.', [], 'Admin.Shipping.Notification');
+                    }
 
-                if (!$carrier->setTaxRulesGroup((int) Tools::getValue('id_tax_rules_group'))) {
-                    $return['has_error'] = true;
-                    $return['errors'][] = $this->trans('An error occurred while saving the tax rules group.', [], 'Admin.Shipping.Notification');
-                }
-
-                if (Tools::getValue('logo')) {
-                    if (Tools::getValue('logo') == 'null' && file_exists(_PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg')) {
-                        unlink(_PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg');
-                    } else {
-                        $logo = basename(Tools::getValue('logo'));
-                        if (!file_exists(_PS_TMP_IMG_DIR_ . $logo) || !copy(_PS_TMP_IMG_DIR_ . $logo, _PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg')) {
+                    if (!$carrier->is_free) {
+                        if (!$this->processRanges((int) $carrier->id)) {
                             $return['has_error'] = true;
-                            $return['errors'][] = $this->trans('An error occurred while saving carrier logo.', [], 'Admin.Shipping.Notification');
+                            $return['errors'][] = $this->trans('An error occurred while saving carrier ranges.', [], 'Admin.Shipping.Notification');
                         }
                     }
+
+                    if (Shop::isFeatureActive() && !$this->updateAssoShop((int) $carrier->id)) {
+                        $return['has_error'] = true;
+                        $return['errors'][] = $this->trans('An error occurred while saving associations of shops.', [], 'Admin.Shipping.Notification');
+                    }
+
+                    if (!$carrier->setTaxRulesGroup((int) Tools::getValue('id_tax_rules_group'))) {
+                        $return['has_error'] = true;
+                        $return['errors'][] = $this->trans('An error occurred while saving the tax rules group.', [], 'Admin.Shipping.Notification');
+                    }
+
+                    if (Tools::getValue('logo')) {
+                        if (Tools::getValue('logo') == 'null' && file_exists(_PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg')) {
+                            unlink(_PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg');
+                        } else {
+                            $logo = basename(Tools::getValue('logo'));
+                            if (!file_exists(_PS_TMP_IMG_DIR_ . $logo) || !copy(_PS_TMP_IMG_DIR_ . $logo, _PS_SHIP_IMG_DIR_ . $carrier->id . '.jpg')) {
+                                $return['has_error'] = true;
+                                $return['errors'][] = $this->trans('An error occurred while saving carrier logo.', [], 'Admin.Shipping.Notification');
+                            }
+                        }
+                    }
+                    $return['id_carrier'] = $carrier->id;
                 }
-                $return['id_carrier'] = $carrier->id;
             }
         }
         die(json_encode($return));
